+++ src/App.less
@@ -33,6 +33,18 @@ body {
   }
 }
 
+.underseksjon {
+  &__header {
+    margin-bottom: 0.5rem;
+  }
+  &__beskrivelse {
+    padding: 0.5rem 0;
+  }
+  &__divider {
+    padding-top: 0.5rem;
+  }
+}
+
 .pagewrapper {
   width: 100%;
 }
