+++ src/components/adresse/AdressePanel.tsx
@@ -9,7 +9,7 @@ interface Props {
 
 const AdressePanel = (props: Props) => (
   <div className="addresse__box">
-    <div className="addresse__tittel">
+    <div className="underseksjon__header">
       <Undertittel>
         <FormattedMessage id={props.tittel} />
       </Undertittel>
