+++ src/components/box/Box.less
@@ -59,7 +59,7 @@
     border: 0.5px solid black;
     width: 100%;
     opacity: 0.5;
-    margin: 1.4rem auto;
+    margin: 1rem auto;
   }
 
   &__content {
@@ -69,7 +69,7 @@
       flex-wrap: wrap;
       list-style-type: none;
       padding-left: 0;
-      margin: 0.5rem 0;
+      margin: 0;
     }
 
     li {
