+++ src/components/kilde/Kilde.less
@@ -3,6 +3,7 @@
     width: 100%;
     display: flex;
     align-items: baseline;
+    margin: 1rem 0;
     @media @mobile {
       flex-direction: column;
     }
@@ -37,6 +38,6 @@
     border: 0.5px solid black;
     width: 100%;
     opacity: 0.5;
-    margin: 1.4rem auto 0.75rem auto;
+    margin: 1rem 0;
   }
 }
