+++ src/index.less
@@ -19,10 +19,10 @@
 @import "./App.less";
 @import "pages/forside/sections/2-brodsmulesti/Brodsmulesti.less";
 @import "pages/forside/sections/3-header/Header.less";
-@import "pages/forside/sections/4-personinfo/adresser/Adresser.less";
-@import "pages/forside/sections/4-personinfo/personalia/Personalia.less";
-@import "pages/forside/sections/4-personinfo/kontakt-informasjon/subsections/DKIF.less";
-@import "pages/forside/sections/4-personinfo/ditt-nav-kontor/DittNavKontor.less";
+@import "pages/forside/sections/4-personinfo/3-adresser/Adresser.less";
+@import "pages/forside/sections/4-personinfo/1-personalia/Personalia.less";
+@import "pages/forside/sections/4-personinfo/2-kontakt-informasjon/subsections/kontakt-og-reservasjonsregisteret/KontaktInfo.less";
+@import "pages/forside/sections/4-personinfo/4-ditt-nav-kontor/DittNavKontor.less";
 @import "pages/forside/sections/5-arbeidsforhold/Arbeidsforhold.less";
 @import "pages/forside/sections/6-eksterne/EksterneLenker.less";
 @import "pages/forside/sections/7-informasjon/MerInformasjon.less";
