+++ src/pages/forside/sections/4-personinfo/2-kontakt-informasjon/subsections/kontakt-og-reservasjonsregisteret/KontaktInfo.less
@@ -2,7 +2,6 @@
   &__overskrift-container {
     display: flex;
     align-items: center;
-    margin-top: 1.5rem;
   }
   &__overskrift-i-icon {
     max-width: 1.5rem;
