+++ src/pages/forside/sections/4-personinfo/3-adresser/Adresser.less
@@ -15,9 +15,7 @@
   &__linje {
     width: 100%;
   }
-  &__midlertidig-adresse {
-    padding-top: 1.5rem;
-  }
+
   &__divider {
     padding: 1rem;
   }
