+++ src/pages/forside/sections/4-personinfo/PersonInfo.tsx
@@ -1,14 +1,14 @@
 import React, { useEffect } from "react";
 import Error, { HTTPError } from "../../../../components/error/Error";
-import Personalia from "./personalia/Personalia";
-import Adresser from "./adresser/Adresser";
-import DittNavKontor from "./ditt-nav-kontor/DittNavKontor";
+import Personalia from "./1-personalia/Personalia";
+import Adresser from "./3-adresser/Adresser";
+import DittNavKontor from "./4-ditt-nav-kontor/DittNavKontor";
 import Spinner from "../../../../components/spinner/Spinner";
 import { PersonInfo } from "../../../../types/personInfo";
 import { fetchPersonInfo } from "../../../../clients/apiClient";
 import { useStore } from "../../../../providers/Provider";
-import KontaktInfo from "./kontakt-informasjon/KontaktInfo";
-import Utbetalinger from "./utbetalinger/Utbetalinger";
+import KontaktInfo from "./2-kontakt-informasjon/KontaktInfo";
+import Utbetalinger from "./5-utbetalinger/Utbetalinger";
 import personaliaIkon from "../../../../assets/img/Personalia.svg";
 import Box from "../../../../components/box/Box";
