--- src/pages/forside/sections/4-personinfo/personalia/Personalia.less
@@ -1,5 +1,0 @@
-.underseksjon {
-  &__beskrivelse {
-    padding: 1rem 0 0 0;
-  }
-}
